refactor: move sync error type to `error.rs`
